feat: split transform chains out from base and add a load function
This was necessary to integrate one test equivalent to resampling with
``antsApplyTransforms``, but via nitransforms.
