+++ nitransforms/io/itk.py
@@ -309,7 +309,7 @@ class ITKCompositeH5:
         except KeyError:
             typo_fallback = "Tranform"
 
-        for xfm in reversed(list(h5group.values())[1:]):
+        for xfm in list(h5group.values())[1:]:
             if xfm["TransformType"][0].startswith(b"AffineTransform"):
                 _params = np.asanyarray(xfm[f"{typo_fallback}Parameters"])
                 xfm_list.append(
